Starbooze fixes
Since food doesn't stack anymore, recipes needed to be adjusted to
require only one of each (at least, one of the items that rot now).
Added reagent tags to the booze items, except for the things I think are
already in the food tab. Might need to go back and add tags to those as
well.