Generic pre and post functions for apiary item spawning. Provide support for using multiple miner-specific frame types in one apiary.
With more than one miner frame type in place, there is the possibility of
generating an ore comb early, thereby suppressing the attempt at generating
the default comb. (This effectively reduces the chance of the default comb,
which is the intent.)

Additionally, each frame type is taken into account, such that if you put an
iron frame and a copper frame in a large apiary, you will get some iron
combs and some copper combs, as well as the default dark earth combs. The
generation ratio should, long-term, be 3:3:4 (6:4 if you put the same type
in both slots) for the large apiary ("alveary").