Rework and fix the honey extractor.
The honey extractor was consuming the wrong amount of honey and would work
with anything in the other input slot without consuming it.

It is now consistent with the honey recipes and now requires jars in the
LOWER input slot, which are consumed as it creates full jars.

It will stop processing if its output slot is full.