Rework the honey jarrer and its interface with the industrial centrifuge.
Rather than relying purely on timing and polling the centrifuge when it
‘should’ have processed enough honey, the honey jarrer now polls the
centrifuge more often for processed honey. The centrifuge keeps track of
what's been processed (of the current type) since it was last asked.

Whenever the jarrer knows that at least three of the same type have been
processed recently, it will produce a jar (if it has jars in stock). If
there is not room to store it, the jar will be dropped on the ground.