This is a spider.  Spiders are now in the mod.  :P  sucks to be you.
Last commit, I paved the way for this mofo, with the spawntables.

This time around, he's here himself, the big man, FRANK!

Adds a monster (Frank) that is spawned with id bighairyfrank. (spawns naturally)

NEXT!

Two new projectiles, the super cool tile sprayer :D :D :D and the webspit.

The tile sprayer places blocks, and is configurable.  Cool.

The web spit is a Frank attack.  Also cool :D

Adds a new block, spidersilkblockmik, that is a web block that can be slowly moved through in all 4 directions.

As part of that, I changed tileEffects.config and one status effect.

Adds the loot table for Frank.  Might add something special later.

And lastly, makes the spidersilkblock not break instantly (it has about the same health as some of the tougher stones now).