honestly, i just forgot to change something
nothing major, but back when i was coding this projectile, before it spawned pass-thru blocks, it trapped effectively by delaying the spawn.
i can now get rid of that delay, making it work better.  also reduced the size of the web blast, to cope with the new center filled area.