Couple of tweaks (perennial plants and fast forwarding)...
Fixed up perennials so that if fertilizer intake causes less seeds to be needed they are returned while if more seeds are needed the growth cycle starts from scratch.

Added a fast forward mechanic which is intentionally lossy so as to keep the server load down.  The mechanic doesn't cost much more than what happened before as it splits the fast forward across updates.  Key variables are at the top of the lua file.