Added dynamic inspection to trays akin to batteries.
The information is based on what is CURRENTLY in the tray slots, not what it's growing.  So you can pick out the right fertilizer and liquid THEN insert your seed of choice.

Also removed racial descriptions from hydroponics trays as that was hiding useful information from players (ie power draw) using the stock races.  Modified description on both trays to remove the explicit number of seeds.