Update 11honeyjarrer.questtemplate
text: changed honey jars to empty bottles, and encouraged the players to place the jarrer next to the centrifuge.  (The empty honey jars were removed, and the usage explanation might clear things up related to unneeded wiring or ITDs.)
completionText: Changed the wording to be less final (there are 2 more quests after this), encouraged the player to experiment with different combs instead of breeding, since that's what quest 12 is for.