Remove Nada
I have made the ''nada'' system obsolete. It has been replaced with a single parameter ''spawnchance'', 0-100. This will allow for easier editing and better control over randomized spawns.

I do apologize for changing this system once again, but this will be the last time. Besides, I think it's worth it, narrowing the spawn chance down to a single parameter instead of having to count "nada"s.