Temporary Dantalion workaround
I don't know what is causing the "not beaming to the Dantalion with a mech" problem. I will try to look into it, but in the mean time, this is a temporary solution.

* Placed an asteroid with a mech summoner station right where the player starts, in the event that they spawn without a mech.

Again, not the most elegant of solutions, but a suitable stopgap until we figure out what is going on here.