Apex Slaver Encounters redone
Those weird Apex Slave encounters have been remade in Tiled. Slaves are now in dedicated buildings or a cave, and are not just wandering about willy-nilly. The 'slaver guild' building isn't made out of packed bones that collapse when farted at anymore. 
Also included "fire" versions of 2 and three which swaps the normal roofing and wood for copper roofing and stone.