Replace incorrect tooltip description of Output2 of pumps
This output is not used anywhere in the current code.
(regardless of what you wire to it, it won't do anything)
However, previous tooltip suggests otherwise, confusing the player.

Note: we are not actually removing Output2 on the Pumps for backward
compatibility (if some player has actually wired it to something,
then having a wire to no-longer-existing output would cause a crash).