Some adjustments to Wrexmspace
I had forgotten to add an anchor that spawns the 'encounter' part of the space encounter with that one, causing you spawn directly behind the ship instead of an asteroid field behind it. 

I never caught it before since it was very difficult if not impossible to test these space encounters natively. In fact it's the first time I've encountered this encounter natively.