Update crewmemberminer.npctype
I expected the featherfall effect to be more powerful than I've found it to be, after a lot of testing. The duration is short enough that it's not consistent enough to rely on, but I don't want to make it a much longer duration because being up constantly would be too strong. Giving them the item magnet II buff in an effort to make them a more useful mining crewmember, will test further and revise if needed.