Altered Alpaca minidungeons
Tightened restrictions on the magicka forest 3 microdungeon to reduce its spawn frequency.

Relaxed restrictions on 1 and 2 microdungeons to allow them to spawn.

Altered the layout of 1 to do so.

Replaced some of the furniture in 2 with more alpaca-themed furniture, e.g. treasure chest -> stone chest.

Added light sources to the microdungeons.