Players too tall
Widened the gap before the reward room of the Verdant Ruins, as players couldn't fit through.
Added a float booster on the railhook section to reduce fall damage taken.
Added loot to the chest at the top of the right path, as it had none specified
Added invisible persistant switches to some of the avian consoles, which are difficult to determine whether they're on or off.