Precursor blast door
Some sliding doors in precursor microdungeons require an explosion to open, but aren't visibly blast doors. They now are.

Also changed the name and blueprint unlocks of the blast door for the sake of the wiki bot, and added a breakDropOptions so that it drops the regular sliding door when mined.