Remove unwanted in-world interactions of Essentia Obscura
1) remove Black Tar + Essentia Obscura = Essentia Obscura
Most likely was forgotten when Essentia Obscura mixings were nerfed.
Essentia Obscura is too lucrative to be mixed in-world at 0 cost.

2) remove Pus + Essentia Obscura - supposed to be a Mixer-only recipe.
(for the same reason, produces Essentia Obscura in Mixer,
which is balanced by speed and power usage of Mixer).

This was producing Corruption block in-world. Removing for consistency
(to not have different results in Mixer and from in-world interaction).