(luacheck) Fix "unused variable" warnings
1. The purpose of this change is to remove all harmless warnings W211,
thus exposing serious warnings that are caused by actual typos/bugs.

2. Details of the fix:

all these variables are unused, so:
a) if they are completely useless, they were removed.
b) if they are used by commented code, they are commented.
c) if they are intended to be used, but are mistakenly marked as local
or are in a wrong scope, then this is fixed.
d) unclear situations are not touched, FIXME comment is added instead.