(luacheck) Remove useless code in Lua of fossil mini-game
Vanilla issue (incomplete code), doesn't really need fixing.

It was supposed to change the volume of sounds when excavating
the fossil (depending on how useful was the hit with the tool?),
but current formula is unusable ("hit" is unknown).

Current behavior defaults to sound 1.0 with 0 loops,
so we leave it at that. Variable "volume" is removed as unused.