Outpost teleport mission
Added a quest to allow players to re-teleport to the Outpost if the bookmark is deleted (or when playing on a world other than the one they started on), as with the Science Outpost. Added it to the gaterepair quest and the fuquickstart/fuadminstart commands.

Removed some unnecessary commands from the quickstart commands.