Remove unused code in monster files
These lines don't do anything, because the value of capturedMonsterType
is the same as this monster's type.

This parameter is used to change the type of monster when it's captured
(e.g. wild Gas Bag -> Gas Bag that doesn't damage blocks),
so it's not needed when resulting type is the same as before.