Make waterbreating work again (e.g. Slimes being able to breathe in Slime)
This feature was previously disabled, because previous implementation was
causing the breath to be lost/regained at double rate
(due to duplicate logic in [stats/player_primary.lua]).

This change solves the problem by not replenishing breath in situations
where vanilla Lua script already does this (either has Breath immunity
or the area is already breathable),
and in all other situations, replenishing it when submerged in liquid.