Improve compatibility with all mods that add dungeons to Garden planet
These mods were incompatible, because their dungeons were competing
with things like "Dark Cavern entrance", causing them not to spawn.

Garden planet was set to have minimum/maximum number of dungeons of 4,
and because FU has exactly 4 dungeons that must spawn there,
all 4 were guaranteed. But some mods were messing this up.

However, as it turns out, we can safely set minumum number of dungeons
to a high number (in this case, 10), because the dungeons are
planet-unique (can't appear more than once),
so if there are only 4 dungeons, and "minumum 10 dungeons" were
requested, then only these 4 will spawn (without duplicates),
which is exactly what we want.

If some mod adds 1 extra dungeon to the Garden planet, then there will
just be 5 dungeons, and FU dungeons won't be prevented to spawn.

The issue can still occur if player installs so many of these mods
that they add 7 or more dungeons, but it's a self-inflicted problem.