fixes to cloud furniture code
several code fixes and image compression.

i scrapped the idea of an animated floating cloud light for now. in order to make it work how i want, we either need to ask thea to use their floating lua animation or we'd need to make our own.

i added those scrapped files in a wip folder in the cloud light object folder and changed the code and filenames so they wouldn't break the replacement cloud light i made in its place.

the rain light works (it turns off and on and has light and sound), but the "on" animation needs to be fixed so it works as intended.

Other than the scrapped custom light and the "working" custom rain light, everything should be working as intended.