QOL and small addition to Erchius Mine
QoL: Room with the infected miner now has the switch to the door. I always thought that loop was rather pointless and annoying. Especially if you take out the infected miner without the other NPC's help. 

Addition: At the end of the map next to the teleporter out, there is now a hidden stash of loot (including a container of erchius loot). This loot is only accessible if you defeat the Erchius Horror the 'hard way' (without using the lasers and without using the microphere bypass)