Bounty Rewards
New bounty rewards that can be purchased from the Peacekeeper shop. 

* Confiscated Guns Crate: Gives 1-5 RNG Guns with a good chance for Neo guns.

* Contraband Ore Box: Gives 60 random ores. Type and tier matches the 'Outlaw Mine' space encounter.

* Illicit Research Datacard: Gives research.

Each has 4 versions (one for tier 2, 3, 4, and 5 of the Peacekeeper outpost).

**NOTES**

All of them currently use the vanilla reward bag as a placeholder sprite/icon, but the file structure is set up for future editing. The balance is not guaranteed and may need to be fine tuned (especially in the research case. It's currently based on 4x the research stick for tier 1.)