Improvements to modularmech.lua for cross-mod compatibility
This patch makes small changes to the Frackin' Universe modularmech.lua
module to expose additional information about mechs to other mods that
are loaded via the Mech Compatibility Core and Loader mods. This will
allow other mods to use this information without the CPU burden of
recalculating it.

* Expose the hazard vulnerabilities as `self.hazardVulnerabilities`.
* Expose the world's environment status effects (e.g., extreme cold,
  heat, etc) as `self.worldEnvironmentStatusEffects`.
* Expose whether the user touched the controls in the current or
  previous tick. This is calculated in the `update` function and exposed
  as `self.hasTouchedControlsRecently`.

Additionally, this patch sets Frackin' Universe to load after Belter
Dungeons to avoid subtle runtime differences depending on the load order
of the two mods.